requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
PyYAML==6.0.1
Werkzeug==3.0.1
zipp==3.18.1
